requirements.txt
streamlit
google-cloud-storage
fastapi
uvicorn

